Revert as these two files are in the winusbfix branch pull request.

+++ src/software-device.h
@@ -117,7 +117,7 @@ namespace librealsense
         friend class software_device;
         stream_profiles _profiles;
         std::map<rs2_frame_metadata_value, rs2_metadata_type> _metadata_map;
-        uint64_t _unique_id;
+        int _unique_id;
 
         class stereo_extension : public depth_stereo_sensor
         {
